+++ app/middleware/checkInitialSetup.server.js
@@ -8,19 +8,18 @@ const checkInitialSetup = async (request) => {
   let { searchParams } = new URL(request.url);
   const shop = searchParams.get("shop");
 
-  const status = await merchantService.merchantStatus(shop);
-  if (status.isRegistered) {
-    // Login
-    console.log("is Registered", status);
-  } else {
-    // Register
-    const { admin } = await authenticate.admin(request);
+  let merchant = await merchantService.getMerchant(shop);
+  
+  if (!merchant) {
+      // Register
+      const { admin } = await authenticate.admin(request);
 
-    const response = await admin.graphql(shopQuery);
-    const responseJson = await response.json();
-
-    initialSetup(responseJson.data.shop);
-  }
+      const response = await admin.graphql(shopQuery);
+      const responseJson = await response.json();
+  
+      merchant = initialSetup(responseJson.data.shop);
+  } 
+  return merchant;
 };
 
 export default checkInitialSetup;