+++ app/services/merchantService.server.js
@@ -1,16 +1,5 @@
 import prisma from "../lib/db.server";
 import logger from "../lib/logger.server";
-import encryption from "../lib/encryption.server";
-import settingsService from "./settingsService.server";
-
-const MERCHANT_ENUM = {
-  STATUS: {
-    ACTIVE: "active",
-    DELETED: "deleted",
-    NO_USER: "no_user",
-    INACTIVE: "inactive",
-  },
-};
 
 const updateMerchant = async (shop, payload) => {
   try {
@@ -57,7 +46,7 @@ const getMerchant = async (shop) => {
 const createMerchant = async (shop) => {
   const payload = {
     shop,
-    status: MERCHANT_ENUM.STATUS.INACTIVE,
+    status: "active",
     subId: "",
     plan: "FREE",
     tagId: "",
@@ -66,26 +55,6 @@ const createMerchant = async (shop) => {
   updateMerchant(shop, payload);
 
   logger.info("User Created on DB", { shop });
-};
-
-const merchantStatus = async (shop) => {
-  const merchant = await getMerchant(shop);
-
-  const status = merchant ? merchant.status : MERCHANT_ENUM.STATUS.NO_USER;
-  const isActive = status === MERCHANT_ENUM.STATUS.ACTIVE;
-  const isInactive = status === MERCHANT_ENUM.STATUS.INACTIVE;
-  const isRegistered = isActive || isInactive;
-  const isDeleted = status === MERCHANT_ENUM.STATUS.DELETED;
-
-  const response = {
-    status,
-    isRegistered,
-    isActive,
-    isInactive,
-    isDeleted,
-  };
-
-  return response;
 };
 
 const deleteMerchant = async (shop) => {
@@ -106,7 +75,7 @@ const merchantService = {
   deleteMerchant,
   merchantStatus,
   createMerchant,
-  updateMerchant,
+  getMerchant,
 };
 
 export default merchantService;