+++ app/utils/initialSetup.server.js
@@ -5,13 +5,15 @@ import shopService from "../services/shopService.server";
 const initialSetup = async (shopData) => {
   const shop = shopData.myshopifyDomain;
   // Create Merchant
-  await merchantService.createMerchant(shop);
+  const merchant = await merchantService.createMerchant(shop);
 
   // Create Settings Data
   await settingsService.createSettings(shop);
 
   // Create Shop
   await shopService.createShop(shopData);
+
+  return merchant;
 };
 
 export default initialSetup;